requirements.txt
python-binance
pandas
numpy
pyTelegramBotAPI
asyncio
python-dotenv
matplotlib